backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
celery>=5.3.4
redis>=5.0.1
httpx>=0.25.0
openai>=1.3.0
geoip2>=4.7.0
python-multipart>=0.0.6
email-validator>=2.0.0